requirements.txt
pandas==2.2.1
minio==7.2.5
sqlalchemy==2.0.28
psycopg2-binary==2.9.9
seaborn==0.13.2
python-dotenv==1.0.1
joblib==1.3.2
dvc==3.48.4
dvc_s3==3.1.0